# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/s3-configs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Returns all S3 backend configurations stored in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "s3-configs"
                ],
                "summary": "List S3 configurations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.S3Configuration"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Creates a new S3 backend configuration. Credentials are encrypted before storage. If marked as default, any existing default is unset.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "s3-configs"
                ],
                "summary": "Create an S3 configuration",
                "parameters": [
                    {
                        "description": "S3 configuration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateS3ConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.S3Configuration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/s3-configs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Returns the details of a specific S3 backend configuration by ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "s3-configs"
                ],
                "summary": "Get an S3 configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "S3 Configuration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.S3Configuration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Updates an existing S3 backend configuration. New credentials are encrypted before storage (credential rotation is always allowed). Changing endpoint, region, bucket_prefix, use_ssl or force_path_style is refused with 409 while any bucket uses the configuration, since it would re-route those buckets' data. bucket_prefix: omit to keep, \"\" to clear. The S3 config cache is invalidated on success.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "s3-configs"
                ],
                "summary": "Update an S3 configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "S3 Configuration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated S3 configuration fields",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateS3ConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.S3Configuration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Admin-only. Deletes an S3 backend configuration. Fails if any buckets are currently using this configuration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "s3-configs"
                ],
                "summary": "Delete an S3 configuration",
                "parameters": [
                    {
                        "type": "string",
                        "description": "S3 Configuration ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Bucket": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_public": {
                    "type": "boolean"
                },
                "lifecycle": {
                    "description": "Lifecycle holds the bucket's lifecycle configuration as JSON (see\nLifecycleConfig); nil = no lifecycle rules.",
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "objects": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Object"
                    }
                },
                "owner": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "owner_id": {
                    "type": "string"
                },
                "quota_bytes": {
                    "description": "QuotaBytes caps the total size of CURRENT objects in the bucket\n(version storage is not counted). 0 = unlimited.",
                    "type": "integer"
                },
                "region": {
                    "type": "string"
                },
                "replicate_to": {
                    "description": "ReplicateTo mirrors this bucket's current objects into another bkt\nbucket (periodic sync; the target is fully managed by replication).",
                    "type": "string"
                },
                "retention_days": {
                    "description": "RetentionDays (WORM-lite): while set, no version of any object can be\npermanently removed until it is older than this many days. Requires\nversioning to be enabled. Plain deletes still create markers (data is\npreserved); version-addressed deletes, lifecycle purges, and bucket\ndeletion are refused for retained data. 0 = off.",
                    "type": "integer"
                },
                "s3_config": {
                    "$ref": "#/definitions/models.S3Configuration"
                },
                "s3_config_id": {
                    "description": "Optional: specific S3 config to use",
                    "type": "string"
                },
                "storage_backend": {
                    "description": "\"local\" or \"s3\"",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "versioning": {
                    "description": "Versioning state: \"\" (never enabled), \"enabled\", or \"suspended\".\nSuspended keeps existing versions browsable but stops creating new ones\n(a documented simplification of AWS's null-version semantics).",
                    "type": "string"
                },
                "webhook_events": {
                    "type": "string"
                },
                "webhook_url": {
                    "description": "Webhook event notifications: POSTed as JSON to WebhookURL for the event\ntypes in WebhookEvents (csv of \"created\",\"removed\"). Empty URL = off.",
                    "type": "string"
                }
            }
        },
...
        "models.CreateS3ConfigRequest": {
            "type": "object",
            "required": [
                "access_key_id",
                "endpoint",
                "name",
                "region",
                "secret_access_key"
            ],
            "properties": {
                "access_key_id": {
                    "type": "string"
                },
                "bucket_prefix": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "force_path_style": {
                    "type": "boolean"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "region": {
                    "type": "string"
                },
                "secret_access_key": {
                    "type": "string"
                },
                "use_ssl": {
                    "type": "boolean"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.S3Configuration": {
            "type": "object",
            "properties": {
                "access_key_id": {
                    "type": "string"
                },
                "bucket_prefix": {
                    "type": "string"
                },
                "buckets": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Bucket"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "force_path_style": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "use_ssl": {
                    "type": "boolean"
                }
            }
        },
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateS3ConfigRequest": {
            "type": "object",
            "properties": {
                "access_key_id": {
                    "type": "string"
                },
                "bucket_prefix": {
                    "description": "BucketPrefix: omitted/null = unchanged, \"\" = clear the prefix.",
                    "type": "string"
                },
                "endpoint": {
                    "type": "string"
                },
                "force_path_style": {
                    "type": "boolean"
                },
                "is_default": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secret_access_key": {
                    "description": "Only update if provided",
                    "type": "string"
                },
                "use_ssl": {
                    "type": "boolean"
                }
            }
        },
...
    }
}